wasm: fix vector abi and allow non-int vectors
Vectors are scalars but can also be passed by reference when their size
is not corresponding to a v128. Therefore, we need to check how vectors
are classified before setting up the return type or parameter type and
not rely on scalarness first. Consistenly with the wasm codegen, vectors
are passed by reference if their size is not equal to 128 or simd
features are not available on the target.